+++ src/types/plugins/server.d.ts
@@ -1,0 +1,6 @@
+export type ServerResponse = {
+    app: any;
+    pouchApp: any;
+    server: any;
+    startupPromise: Promise<void>;
+}